+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Watermark Changelog
 
+### 1.0.1 - 2020-07-28
+
+- Added position of watermark.
+- Added transparency level for watermark.
+
 ### 1.0.0 - 2020-06-04
 
 - First release.